requirements.txt
# Python Dependencies for Marine Ministry Crawler
# Install: pip install -r requirements.txt

# Web scraping
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.2

# Data processing
pandas==2.3.3
pytz==2025.2
openpyxl==3.1.5

# Google Sheets API
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
